requirements.txt
Django==1.9
Pillow==3.0.0
pytz==2015.7
wheel==0.24.0
audiotools==3.1a1
fusepy==2.0.2
